+++ types/protocol-mapping.d.ts
@@ -819,6 +819,11 @@ export namespace ProtocolMapping {
          * or a command below.
          */
         'FedCm.dialogClosed': [Protocol.FedCm.DialogClosedEvent];
+        /**
+         * Event for when a GATT operation of |type| to the peripheral with |address|
+         * happened.
+         */
+        'BluetoothEmulation.gattOperationReceived': [Protocol.BluetoothEmulation.GattOperationReceivedEvent];
     }
 
     export interface Commands {
@@ -5295,6 +5300,15 @@ export namespace ProtocolMapping {
          */
         'BluetoothEmulation.simulateAdvertisement': {
             paramsType: [Protocol.BluetoothEmulation.SimulateAdvertisementRequest];
+            returnType: void;
+        };
+        /**
+         * Simulates the response code from the peripheral with |address| for a
+         * GATT operation of |type|. The |code| value follows the HCI Error Codes from
+         * Bluetooth Core Specification Vol 2 Part D 1.3 List Of Error Codes.
+         */
+        'BluetoothEmulation.simulateGATTOperationResponse': {
+            paramsType: [Protocol.BluetoothEmulation.SimulateGATTOperationResponseRequest];
             returnType: void;
         };
     }